hotfix: Change user model kakao_id column type
카카오 id 범위에 맞게 int -> bigint로 모델 수정

+++ backend/app/app/models/user.py
@@ -1,6 +1,6 @@
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
+from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, BigInteger
 from sqlalchemy.orm import relationship
 
 from app.db.base_class import Base
@@ -14,7 +14,7 @@ class User(Base):
     id = Column(String, primary_key=True, index=True)
     email = Column(String, unique=True, index=True, nullable=True)
     hashed_password = Column(String, nullable=True)
-    kakao_id = Column(Integer, nullable=True)
+    kakao_id = Column(BigInteger, nullable=True)
     color_id = Column(Integer, ForeignKey("color.id"))
     nickname = Column(String, nullable=True)
     is_active = Column(Boolean(), default=False)
